web/schema.sql: add indexes on the database schema
This commit adds indexes on the two columns of `symbol_per_result` to
significantly improve queries involving symbols (this is an over 80M
rows table). This table is heavily used to retrieve results matching a
symbol, and without those indexes, the duration of such queries is
prohibitive.

While at it, we also mark these columns as foreign keys to ensure the
integrity of the database.

Signed-off-by: Victor Huesca <victor.huesca@bootlin.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>