scripts/autobuild-run: create instance directory before using it
In commit 01f624d343efee85999d9f2bc225d04c1593aa69 ("autobuild-run:
move log variable to Builder class"), setting the self.log variable
was moved to the constructor of the Builder class. This opens the
"instance.log" in the instance directory. However, the instance
directory is created later, in the run_instance() method. So since
this commit, if the instance directory doesn't already exist, the
script crashes with an exception.

Fix this by moving the instance directory creation inside the Builder
class constructor.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>