utils/daily-mail: remove the unnecessary 'get_notification_for_dev' function
Python dictionaries provide a way to return a value associated to a key
while adding the corresponding entry in case the key doesn't exists yet
(it acts like a default_dict from itertools).
Since the 'get_notification_for_dev' does exactly that, this patch
remove this function and replace it by the build-in
`dict.setdefault(key, value)`.

Signed-off-by: Victor Huesca <victor.huesca@bootlin.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>