utils/daily-mail: replace '%' style format strings by the newer '{}' version
Python provides several ways to perform string formatting. The 1st is C
style inspired and use the '%' operator, this is the original way to
create format strings in python. The second way -- which is the standard
in python3 -- is the '.format()' method, this version is more flexible,
removes the strong type constraints and allow more format styles. This
method is also noted as the preferred way over the '%' formatting [1].

In order to introduce new kind of notification, this patch update the
current format strings to this new standard.

[1] https://docs.python.org/2/library/stdtypes.html#str.format

Signed-off-by: Victor Huesca <victor.huesca@bootlin.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>