autobuild-run: make diffoscope output a JSON-formatted file as well
Normal diffoscope output is readable by humans but not really
convenient when working with it in code. JSON can be easily
read and extracted information from.

We still keep the normal text version because it's easy to parse by a
human, but we limit its size to 40KB.

While at it, rename the files to more sensible names:
diffoscope-results.txt and diffoscope-results.json.

Signed-off-by: Atharva Lele <itsatharva@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>