web/db.inc.php: add support to test if the database supports a feature
This patch allows to test if the database supports a given feature.

Any feature to be test against can be added by simply adding a new case
to the switch.

This patch allows to check if the database engine supports the
'intersect' SQL keyword. This keyword was implemented in mariadb 10.3.10
released in oct. 2018

Signed-off-by: Victor Huesca <victor.huesca@bootlin.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>