web/funcs.inc.php: add support for filtering autobuild results by configuration options
This patch improves 'bab_get_results()' and 'bab_total_results_count()'
so that we can filter autobuilder results from the database with a list
of configuration options and values.

The SQL query has been optimized to scale the better when multiple
symbols are involved altogether. There is at least 3 ways to get
the same results: 'join', 'where in' and 'intersect'.
- The 'join' does not scale at all and can only return a dozen
  configurations in a reasonable amount of time.
- The 'intersect' outperforms the two others and allows to add as many
  configuration symbols as wanted, even when thousand of configs are
  returned.
- The 'where in' is the compromise solution. It works well for asking
  two or three configuration symbols at the same time but can become a
  real bottleneck when those symbols are commons are returns a too many
  configurations.
  Also its results depends on the underlying DBMS, this make its
  performances unpredictable.

The current deployment uses an old mysql server that does not supports
the 'intersect' queries. In order to scale well with a potential
migration of the SQL server to a MariaDB 10.3.10+, this patch implements
both the 'intersect' and 'where in' versions of the query. The default
behavior is to look for 'intersect' support and fallback on the 'where
in' query otherwise.

Signed-off-by: Victor Huesca <victor.huesca@bootlin.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>