web/search: new page to help making request to the autobuild database
The current way to make searches is to manually pass GET parameters to
the URL, which is nice for automation and simple request but it gets
messy with more complicated requests -- especially with the
introduction of configuration symbols.

This patch provides a handy interface to ask more sophisticated configs
to the autobuild database.

The implementation uses normal HTML forms but implements a hook to the
submit method in order to handle config symbols from the textarea and
keep the URL clean.

Signed-off-by: Victor Huesca <victor.huesca@bootlin.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>