utils/daily-mail: add support for fetching results of Gitlab CI jobs
Buildroot has a mirror repository on gitlab.com that is used to run
continuous integration tasks. Among these tasks, a complete check of
defconfigs and runtime-tests is regularly performed.

This patch allows to retrieve tests that have failed over a given
period of time. It is based on the (very imperfect) gitlab-ci API.

Due to multiple limitations of the API (pagination of results, crucial
data requiring a dedicated request, etc.), this implementation has
several optimizations to be able to retrieve the needed information.

Signed-off-by: Victor Huesca <victor.huesca@bootlin.com>
[Thomas:
 - filter the pipelines by branch earlier (before getting the pipeline
   details)
 - filter the pipelines by status, keeping only failed pipelines
 - rename date_from to 'ndays', as it's really a number of days, not a
   date
 - fix typos]
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>