scripts/autobuild-run: calculate dldir as an absolute path
Right now, self.dldir is a relative path, equal to
"instance-X/dl". Then, the Buildroot build is executed with
BR2_DL_DIR=self.dldir, which means that self.dldir is interpreted
relatively to the Buildroot source directory.

Due to this, we get the following directory organization:

 instance-0
  + buildroot
    + instance-0
      + dl
  + instance.log
  + output

Instead of the excepted organization:

 instance-0
 + buildroot
 + dl
 + instance.log
 + output

We fix this issue (which was introduced when the Builder class was
created) by making self.dldir an absolute path.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Reviewed-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Reviewed-by: Atharva Lele <itsatharva@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>