scripts/autobuild-run: properly delete output dir even with write-protected directories
We're already deleting the output directory before a new build with
"rm -rf" so that write protected files are removed as well, which
shutil.rmtree() doesn't do.

However "rm -rf" fails on write protected directories.

So instead, switch back to using shutil.rmtree(), with an error handling
function that takes care of setting appropriate permissions to allow the
removal:

  - set the parent directory as +rwx so we can remove the entry,

  - set files as +rw and direcotries as _rwx; since +x on files is not a
    problem, simplify the logic and set +rwx on the entity to remove

This solves an autobuild-run failure that occurs when the output
directory cannot be deleted due to write-protected directories. This
is happening with mender-artefact Go modules for example.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
[yann.morin.1998@free.fr:
  - simplify the error handler
  - update commit log accordingly
  - s/folder/directory/
  - add simple docstring
]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>