utils/daily-mail: rework the global e-mail format
We want to take into account all Gitlab CI results, not just the ones
for the explicit branches we are tracking. Indeed, when a Gitlab CI
pipeline runs on a Buildroot tag, it is this tag that is used as the
"ref" in the pipeline results, and it does not appear to belong to
master, next, or some other branch.

This commit changes how the global-email is formatted. After this
commit, it looks like this:

Autobuild statistics for 2020-08-09
===================================

<overall stats here>

Classification of failures by reason for master
-----------------------------------------------

<classification of failures>

Detail of failures for master
-----------------------------

<list of failures>

Classification of failures by reason for next
-----------------------------------------------

<classification of failures>

Detail of failures for next
-----------------------------

<list of failures>

Gitlab CI results for 2020-08-09
================================

Detail of defconfig failures for 2020.08-rc1
--------------------------------------------

<defconfig failures>

Detail of runtime-test failures for 2020.08-rc1
-----------------------------------------------

<runtime test failures>

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>