utils/daily-mail: rewrite the developers_email() notifications
Just like we did for the global e-mail, the per-developer e-mail also
needs to be modified to report all Gitlab CI failures, and not just
the ones for the list of branches we are explicitly tracking.

This commit therefore creates two sections in the mail, one for the
autobuilder results, and one for the Gitlab CI results.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>