utils/daily-mail: handle Gitlab CI pipelines from older branches
In the newer Buildroot branches, we have this new Gitlab CI way of
describing the CI jobs, which results in having the jobs appearing in
a "downstream" pipeline. In commit
82a8886d33c1ca11486f59702ba83c67432330e5, daily-mail was changed to
accommodate those changes, but forgot that Gitlab CI pipelines
happening on older branches (such as 2020.02.x) would stick to the
previous way, without downstream pipelines.

Due to this, daily-mail was aborting with an exception when a Gitlab
CI on 2020.02.x had happened. This commit adjusts daily-mail to
account for both situations.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>