utils/daily-mail: only look at pipelines that have source == trigger
The only pipelines that are relevant for this script are the pipelines
that do the testing of defconfigs and runtime tests. Such pipelines
are always explicitly triggered (as opposed by implicitly triggered by
doing a push). We can reduce the number of pipelines we look at by
filtering the pipelines that have "source": "trigger".

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>