scripts/autobuild-run: remove whole directories in $(DL_DIR)
Prior to running a build, we remove 5 random files from $(DL_DIR) in
order to regularly forces our autobuilder instances to redownload
files rather than always rely on their local $(DL_DIR).

The logic was smart enough to skip .git/ folders, but with the
addition of br-cargo-home/ and br-flutter-pub-cache/ under $(DL_DIR),
we now have many more individual files, and it does not make sense to
remove individual files.

So we simplify the logic and simply select 5 random sub-directories of
$(DL_DIR) and remove them. So it will remove entire package download
directories, or potentially remove the Cargo cache or the Flutter
cache.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>